UPDATE: Move autofill into own module
